Migrate to new class name "QuiverTwitterFollower", obsolete "QuiverQuantTwitterFollowers"
